requirements.txt
# Core CRM System Dependencies
pandas==2.1.4
openpyxl==3.1.2
streamlit==1.29.0
plotly==5.17.0

# AI and Advanced Features
openai==1.3.7
python-dotenv==1.0.0

# Data Processing and Export
numpy==1.24.3
python-dateutil==2.8.2
xlsxwriter
xlwt

# Google Sheets Integration (Future)
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Web Deployment and Security
streamlit-authenticator
cryptography

# Database and Cloud Storage
sqlalchemy
streamlit-option-menu
